requirements.txt
discord.py[voice]
python-dotenv
youtube-dl
matplotlib
numpy
pandas
requests
easy-pil
ffmpeg